LPS-63470 move classes to definition-api
