+++ modules/apps/business-productivity/portal-workflow/portal-workflow-kaleo-definition-api/src/main/java/com/liferay/portal/workflow/kaleo/definition/JoinXor.java
